Switch MultiInterfaceController to use variadic templates.
